revision of loader refactor
