+++ conans/client/export.py
@@ -10,10 +10,39 @@ from conans.errors import ConanException
 from conans.model.manifest import FileTreeManifest
 from conans.client.output import ScopedOutput
 from conans.client.file_copier import FileCopier
+from conans.model.conan_file import create_exports, create_exports_sources
+from conans.client.loader_parse import load_conanfile_class
 
 
-def export_conanfile(output, paths, conanfile, origin_folder, conan_ref, short_paths,
-                     keep_source):
+def load_export_conanfile(conanfile_path, output):
+    conanfile = load_conanfile_class(conanfile_path)
+
+    for field in ["url", "license", "description"]:
+        field_value = getattr(conanfile, field, None)
+        if not field_value:
+            output.warn("Conanfile doesn't have '%s'.\n"
+                        "It is recommended to add it as attribute" % field)
+    if getattr(conanfile, "conan_info", None):
+        output.warn("conan_info() method is deprecated, use package_id() instead")
+
+    try:
+        # Exports is the only object field, we need to do this, because conan export needs it
+        conanfile.exports = create_exports(conanfile)
+        conanfile.exports_sources = create_exports_sources(conanfile)
+    except Exception as e:  # re-raise with file name
+        raise ConanException("%s: %s" % (conanfile_path, str(e)))
+
+    # check name and version were specified
+
+    if not hasattr(conanfile, "name") or not conanfile.name:
+        raise ConanException("conanfile didn't specify name")
+    if not hasattr(conanfile, "version") or not conanfile.version:
+        raise ConanException("conanfile didn't specify version")
+
+    return conanfile
+
+
+def export_conanfile(output, paths, conanfile, origin_folder, conan_ref, keep_source):
     destination_folder = paths.export(conan_ref)
     previous_digest = _init_export_folder(destination_folder)
     execute_export(conanfile, origin_folder, destination_folder, output)
@@ -30,7 +59,7 @@ def export_conanfile(output, paths, conanfile, origin_folder, conan_ref, short_p
         output.info('Folder: %s' % destination_folder)
         modified_recipe = True
 
-    source = paths.source(conan_ref, short_paths)
+    source = paths.source(conan_ref, conanfile.short_paths)
     dirty = os.path.join(source, DIRTY_FILE)
     remove = False
     if os.path.exists(dirty):
