+++ conans/client/loader.py
@@ -1,25 +1,16 @@
-import imp
-import inspect
 import os
-import sys
-import uuid
 
-from conans.client.generators import _save_generator
 from conans.errors import ConanException, NotFoundException
-from conans.model.conan_file import ConanFile, create_exports, create_exports_sources
-from conans.model.conan_generator import Generator
+from conans.model.conan_file import ConanFile
 from conans.model.options import OptionsValues
 from conans.model.ref import ConanFileReference
-from conans.model.scope import Scopes
 from conans.model.settings import Settings
 from conans.model.values import Values
-from conans.util.config_parser import ConfigParser
 from conans.util.files import load
-from conans.util.files import rmdir
-from conans.paths import BUILD_INFO, CONANINFO
+from conans.paths import BUILD_INFO
 from conans.model.build_info import DepsCppInfo
 from conans.model.profile import Profile
-from conans.model.info import ConanInfo
+from conans.client.loader_parse import ConanFileTextLoader, load_conanfile_class
 
 
 def _apply_initial_deps_infos_to_conanfile(conanfile, initial_deps_infos):
@@ -57,143 +48,21 @@ def _load_info_file(current_path, conanfile, output, error=False):
         raise ConanException("Parse error in '%s' file in %s" % (BUILD_INFO, current_path))
 
 
-def _parse_module(conanfile_module, filename):
-    """ Parses a python in-memory module, to extract the classes, mainly the main
-    class defining the Recipe, but also process possible existing generators
-    @param conanfile_module: the module to be processed
-    @param consumer: if this is a root node in the hierarchy, the consumer project
-    @return: the main ConanFile class from the module
-    """
-    result = None
-    for name, attr in conanfile_module.__dict__.items():
-        if "_" in name:
-            continue
-        if (inspect.isclass(attr) and issubclass(attr, ConanFile) and attr != ConanFile and
-                attr.__dict__["__module__"] == filename):
-            if result is None:
-                result = attr
-            else:
-                raise ConanException("More than 1 conanfile in the file")
-        if (inspect.isclass(attr) and issubclass(attr, Generator) and attr != Generator and
-                attr.__dict__["__module__"] == filename):
-                _save_generator(attr.__name__, attr)
-
-    if result is None:
-        raise ConanException("No subclass of ConanFile")
-
-    return result
-
-
-def _parse_file(conan_file_path):
-    """ From a given path, obtain the in memory python import module
-    """
-    # Check if precompiled exist, delete it
-    if os.path.exists(conan_file_path + "c"):
-        os.unlink(conan_file_path + "c")
-
-    # Python 3
-    pycache = os.path.join(os.path.dirname(conan_file_path), "__pycache__")
-    if os.path.exists(pycache):
-        rmdir(pycache)
-
-    if not os.path.exists(conan_file_path):
-        raise NotFoundException("%s not found!" % conan_file_path)
-
-    filename = os.path.splitext(os.path.basename(conan_file_path))[0]
-
-    try:
-        current_dir = os.path.dirname(conan_file_path)
-        sys.path.append(current_dir)
-        old_modules = list(sys.modules.keys())
-        loaded = imp.load_source(filename, conan_file_path)
-        # Put all imported files under a new package name
-        module_id = uuid.uuid1()
-        added_modules = set(sys.modules).difference(old_modules)
-        for added in added_modules:
-            module = sys.modules[added]
-            if module:
-                folder = os.path.dirname(module.__file__)
-                if folder.startswith(current_dir):
-                    module = sys.modules.pop(added)
-                    sys.modules["%s.%s" % (module_id, added)] = module
-    except Exception:
-        import traceback
-        trace = traceback.format_exc().split('\n')
-        raise ConanException("Unable to load conanfile in %s\n%s" % (conan_file_path,
-                                                                     '\n'.join(trace[3:])))
-    finally:
-        sys.path.pop()
-
-    return loaded, filename
-
-
-def load_conanfile_class(conanfile_path, check_name_version=False):
-    """ Load only the class of the ConanFile recipe, but do not instantiate the object
-    It is needed for the 'conan export' command
-    """
-    loaded, filename = _parse_file(conanfile_path)
-    try:
-        result = _parse_module(loaded, filename)
-        # Exports is the only object field, we need to do this, because conan export needs it
-        result.exports = create_exports(result)
-        result.exports_sources = create_exports_sources(result)
-    except Exception as e:  # re-raise with file name
-        raise ConanException("%s: %s" % (conanfile_path, str(e)))
-
-    # check name and version were specified
-    if check_name_version:
-        if not hasattr(result, "name") or not result.name:
-            raise ConanException("conanfile didn't specify name")
-        if not hasattr(result, "version") or not result.version:
-            raise ConanException("conanfile didn't specify version")
-
-    return result
-
-
-def _loader(runner, settings, profile):
-    return ConanFileLoader(runner,
-                           settings=settings,
-                           package_settings=profile.package_settings_values,
-                           options=profile.options, scopes=profile.scopes,
-                           env_values=profile.env_values)
-
-
-def _get_single_loader(current_path, settings, runner):
-    profile = Profile()
-    conan_info_path = os.path.join(current_path, CONANINFO)
-    if conan_info_path and os.path.exists(conan_info_path):
-        existing_info = ConanInfo.load_file(conan_info_path)
-        settings.values = existing_info.full_settings
-        profile.options = existing_info.full_options
-        profile.scopes = existing_info.scope
-        profile.env_values = existing_info.env_values
-
-    return _loader(runner, settings, profile)
-
-
-def install_loader(settings, profile, runner):
-    settings.values = profile.settings_values
-    return _loader(runner, settings, profile)
-
-
-def load_conanfile_single(conanfile_path, current_path, settings, runner, output, reference=None,
-                          error=False):
-    loader = _get_single_loader(current_path, settings, runner)
-    consumer = not reference
-    conanfile = loader.load_conan(conanfile_path, output, consumer, reference)
-    _load_info_file(current_path, conanfile, output, error)
-    return conanfile
-
-
-def load_conanfile_txt_single(conanfile_path, current_path, settings, runner, output, error=False):
-    loader = _get_single_loader(current_path, settings, runner)
-    conanfile = loader.load_conan_txt(conanfile_path, output)
+def load_consumer_conanfile(conanfile_path, current_path, settings, runner, output, reference=None,
+                            error=False):
+    profile = Profile.read_conaninfo(current_path)
+    loader = ConanFileLoader(runner, settings, profile)
+    if conanfile_path.endswith(".py"):
+        consumer = not reference
+        conanfile = loader.load_conan(conanfile_path, output, consumer, reference)
+    else:
+        conanfile = loader.load_conan_txt(conanfile_path, output)
     _load_info_file(current_path, conanfile, output, error)
     return conanfile
 
 
 class ConanFileLoader(object):
-    def __init__(self, runner, settings, package_settings, options, scopes, env_values):
+    def __init__(self, runner, settings, profile):
         '''
         @param settings: Settings object, to assign to ConanFile at load time
         @param options: OptionsValues, necessary so the base conanfile loads the options
@@ -202,24 +71,22 @@ class ConanFileLoader(object):
         @param cached_env_values: EnvValues object
         '''
         self._runner = runner
+        settings.values = profile.settings_values
         assert settings is None or isinstance(settings, Settings)
-        assert options is None or isinstance(options, OptionsValues)
-        assert scopes is None or isinstance(scopes, Scopes)
         # assert package_settings is None or isinstance(package_settings, dict)
         self._settings = settings
-        self._user_options = options
-        self._scopes = scopes
+        self._user_options = profile.options
+        self._scopes = profile.scopes
 
-        self._package_settings = package_settings
-        self._env_values = env_values
+        self._package_settings = profile.package_settings_values
+        self._env_values = profile.env_values
         self.initial_deps_infos = None
 
     def load_conan(self, conanfile_path, output, consumer=False, reference=None):
         """ loads a ConanFile object from the given file
         """
-        loaded, filename = _parse_file(conanfile_path)
+        result = load_conanfile_class(conanfile_path)
         try:
-            result = _parse_module(loaded, filename)
             # Prepare the settings for the loaded conanfile
             # Mixing the global settings with the specified for that name if exist
             tmp_settings = self._settings.copy()
@@ -307,55 +174,3 @@ class ConanFileLoader(object):
         conanfile.scope = self._scopes.package_scope()
 
         return conanfile
-
-
-class ConanFileTextLoader(object):
-    """Parse a plain requirements file"""
-
-    def __init__(self, input_text):
-        # Prefer composition over inheritance, the __getattr__ was breaking things
-        self._config_parser = ConfigParser(input_text,  ["requires", "generators", "options",
-                                                         "imports"], parse_lines=True)
-
-    @property
-    def requirements(self):
-        """returns a list of requires
-        EX:  "OpenCV/2.4.10@phil/stable"
-        """
-        return [r.strip() for r in self._config_parser.requires.splitlines()]
-
-    @property
-    def options(self):
-        return self._config_parser.options
-
-    @property
-    def import_parameters(self):
-        ret = []
-        local_install_text = self._config_parser.imports
-        for local_install_line in local_install_text.splitlines():
-            invalid_line_msg = "Invalid imports line: %s" \
-                               "\nEX: OpenCV/lib, * -> ./lib" % local_install_line
-            try:
-                if local_install_line.startswith("/") or local_install_line.startswith(".."):
-                    raise ConanException("Import's paths can't begin with '/' or '..'")
-                pair = local_install_line.split("->")
-                source = pair[0].strip().split(',', 1)
-                dest = pair[1].strip()
-                src, pattern = source[0].strip(), source[1].strip()
-                ret.append((pattern, dest, src))
-            except ConanException as excp:
-                raise ConanException("%s\n%s" % (invalid_line_msg, excp.message))
-            except:
-                raise ConanException(invalid_line_msg)
-        return ret
-
-    @property
-    def generators(self):
-        return self._config_parser.generators.splitlines()
-
-    @staticmethod
-    def imports_method(conan_file, parameters):
-        def imports():
-            for import_params in parameters:
-                conan_file.copy(*import_params)
-        return imports
