+++ conans/test/functional/conanfile_loader_test.py
@@ -9,13 +9,14 @@ from mock import Mock
 from conans.model.settings import Settings
 from conans.test.utils.test_files import temp_folder
 from conans.model.scope import Scopes
+from conans.model.profile import Profile
+from collections import OrderedDict
 
 
 class ConanLoaderTest(unittest.TestCase):
 
     def requires_init_test(self):
-        loader = ConanFileLoader(None, Settings(), None, OptionsValues.loads(""), Scopes(),
-                                 None)
+        loader = ConanFileLoader(None, Settings(), Profile())
         tmp_dir = temp_folder()
         conanfile_path = os.path.join(tmp_dir, "conanfile.py")
         conanfile = """from conans import ConanFile
@@ -82,7 +83,7 @@ OpenCV2:other_option=Cosa
         tmp_dir = temp_folder()
         file_path = os.path.join(tmp_dir, "file.txt")
         save(file_path, file_content)
-        loader = ConanFileLoader(None, Settings(), None, OptionsValues.loads(""), Scopes(), None)
+        loader = ConanFileLoader(None, Settings(), Profile())
         ret = loader.load_conan_txt(file_path, None)
         options1 = OptionsValues.loads("""OpenCV:use_python=True
 OpenCV:other_option=False
@@ -109,8 +110,7 @@ OpenCV/2.4.104phil/stable <- use_python:True, other_option:False
         tmp_dir = temp_folder()
         file_path = os.path.join(tmp_dir, "file.txt")
         save(file_path, file_content)
-        loader = ConanFileLoader(None, Settings(), None, OptionsValues.loads(""),
-                                 Scopes(), None)
+        loader = ConanFileLoader(None, Settings(), Profile())
         with self.assertRaisesRegexp(ConanException, "Wrong package recipe reference(.*)"):
             loader.load_conan_txt(file_path, None)
 
@@ -122,8 +122,7 @@ OpenCV/bin/* - ./bin
         tmp_dir = temp_folder()
         file_path = os.path.join(tmp_dir, "file.txt")
         save(file_path, file_content)
-        loader = ConanFileLoader(None, Settings(), None, OptionsValues.loads(""),
-                                 Scopes(), None)
+        loader = ConanFileLoader(None, Settings(), Profile())
         with self.assertRaisesRegexp(ConanException, "is too long. Valid names must contain"):
             loader.load_conan_txt(file_path, None)
 
@@ -141,28 +140,23 @@ class MyTest(ConanFile):
         save(conanfile_path, conanfile)
 
         # Apply windows for MyPackage
-        package_settings = {"MyPackage": [("os", "Windows")]}
-        loader = ConanFileLoader(None, Settings({"os": ["Windows", "Linux"]}),
-                                 package_settings, OptionsValues.loads(""), Scopes(),
-                                 None)
+        profile = Profile()
+        profile.package_settings = {"MyPackage": OrderedDict([("os", "Windows")])}
+        loader = ConanFileLoader(None, Settings({"os": ["Windows", "Linux"]}), profile)
 
         recipe = loader.load_conan(conanfile_path, None)
         self.assertEquals(recipe.settings.os, "Windows")
 
         # Apply Linux for MyPackage
-        package_settings = {"MyPackage": [("os", "Linux")]}
-        loader = ConanFileLoader(None, Settings({"os": ["Windows", "Linux"]}),
-                                 package_settings, OptionsValues.loads(""), Scopes(),
-                                 None)
+        profile.package_settings = {"MyPackage": OrderedDict([("os", "Linux")])}
+        loader = ConanFileLoader(None, Settings({"os": ["Windows", "Linux"]}), profile)
 
         recipe = loader.load_conan(conanfile_path, None)
         self.assertEquals(recipe.settings.os, "Linux")
 
         # If the package name is different from the conanfile one, it wont apply
-        package_settings = {"OtherPACKAGE": [("os", "Linux")]}
-        loader = ConanFileLoader(None, Settings({"os": ["Windows", "Linux"]}),
-                                 package_settings, OptionsValues.loads(""), Scopes(),
-                                 None)
+        profile.package_settings = {"OtherPACKAGE": OrderedDict([("os", "Linux")])}
+        loader = ConanFileLoader(None, Settings({"os": ["Windows", "Linux"]}), profile)
 
         recipe = loader.load_conan(conanfile_path, None)
         self.assertIsNone(recipe.settings.os.value)
